# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LocalPro API — hyperlocal services marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a completed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /providers/{id}/reviews:
    get:
      operationId: listReviewsForProvider
      tags: [reviews]
      summary: List reviews for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # ── DISPUTES ──────────────────────────────────────────────────────────────
...
  /upskilling/modules:
    get:
      operationId: listUpskillingModules
      tags: [upskilling]
      summary: List upskilling modules
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UpskillingModule"
  /upskilling/modules/{id}:
    get:
      operationId: getUpskillingModule
      tags: [upskilling]
      summary: Get upskilling module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpskillingModule"
  /upskilling/modules/{id}/complete:
    post:
      operationId: completeUpskillingModule
      tags: [upskilling]
      summary: Mark a module as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderProgress"
...
components:
  schemas:
...
    Review:
      type: object
      required: [id, bookingId, rating, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        flagged:
          type: boolean
        createdAt:
          type: string
        residentName:
          type: ["string", "null"]
    ReviewInput:
      type: object
      required: [bookingId, rating]
      properties:
        bookingId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]

    # ── Dispute schemas ────────────────────────────────────────────────────
...
    UpskillingModule:
      type: object
      required: [id, title, category]
      properties:
        id:
          type: integer
        title:
          type: string
        contentUrl:
          type: ["string", "null"]
        category:
          type: string
        completionCriteria:
          type: ["string", "null"]
        completed:
          type: boolean
    ProviderProgress:
      type: object
      required: [moduleId, completed]
      properties:
        moduleId:
          type: integer
        completed:
          type: boolean
        badgeUnlocked:
          type: boolean
...